Store a mask; Much smaller memory footprint (#70)
Rather than storing a list of zero indices, store a zero mask.

With this change, the SparseWeightsBase no longer prescribes how the
zeroes should be stored; the subclass gets to decide this. This
enables subclasses to choose more efficient case-specific rezeroing
techniques, for example when sparsity is structured (e.g.
channel-to-channel).

Memory usage of resnet50, before and after:

Before:
672,979,068 bytes total
570,523,648 bytes of SparseWeights zero indices

After:
152,778,876 bytes total
 50,323,456 bytes of SparseWeights masks

And the memory improvement gets considerably better if the sparsity is
structured (e.g. channel to channel), enabling broadcasting.