KWinners Perf: Make break_ties=True faster, use less memory
In a resnet50 experiment with break_ties=True, this reduces the
KWinners overhead by about 38% (to 62% of its current value)

Saving the index list uses a lot of memory. Using masked_fill is more
efficient and has a fixed memory cost rather than scaling with
density. A resnet50 experiment was running out of memory, and this
change solves that problem.

This change enables us to incorporate ReLU, which adds an extra
speedup.

Another speedup comes from using sorted=False on topk.

This change eliminates the need to have our own torch.autograd
functions. It would already be possible to get rid of these anyway;
`tensor.scatter_` has all the gradient functionality built in that we
need. But using that solution doesn't solve the memory issue and is
slower (because it requires a `tensor.gather`, and because it doesn't
give us a good way to incoprorate ReLU)